+++ include/sentry.h
@@ -1509,6 +1509,15 @@ SENTRY_API void sentry_set_fingerprint_n(
 SENTRY_API void sentry_remove_fingerprint(void);
 
 /**
+ * Set the trace. The primary use for this is to allow other SDKs to propagate
+ * their trace context to connect events on all layers
+ */
+SENTRY_API void sentry_set_trace(
+    const char *trace_id, const char *parent_span_id);
+SENTRY_API void sentry_set_trace_n(const char *trace_id, size_t trace_id_len,
+    const char *parent_span_id, size_t parent_span_id_len);
+
+/**
  * Sets the transaction.
  */
 SENTRY_API void sentry_set_transaction(const char *transaction);